refactor: 이미지 처리를 위한 별도 서비스를 추가
