+++ src/lib/aws-sdk/s3/aws-s3.interface.ts
@@ -1,14 +1,3 @@
-export interface ResizeImageToWebpOptions {
-  /** Buffer 이미지 */
-  image: Buffer;
-
-  /** 너비 */
-  width: number;
-
-  /** 높이 */
-  height?: number;
-}
-
 export interface GetFileFromS3Options {
   /** 버킷 이름 */
   bucket: string;
