+++ src/lib/aws-sdk/s3/aws-s3.module.ts
@@ -2,6 +2,8 @@ import { S3 } from '@aws-sdk/client-s3';
 import { Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 
+import { AwsS3Service } from './aws-s3.service';
+
 @Module({
   imports: [],
   providers: [
@@ -19,5 +21,6 @@ import { ConfigService } from '@nestjs/config';
       inject: [ConfigService],
     },
   ],
+  exports: [AwsS3Service],
 })
 export class AwsS3Module {}
