+++ src/lib/image-processing/image-processing.interface.ts
@@ -1,0 +1,10 @@
+export interface ResizeImageToWebpOptions {
+  /** Buffer 이미지 */
+  image: Buffer;
+
+  /** 너비 */
+  width: number;
+
+  /** 높이 */
+  height?: number;
+}
