+++ src/lib/image-processing/image-processing.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+
+import { ImageProcessingService } from './image-processing.service';
+
+@Module({
+  providers: [ImageProcessingService],
+  exports: [ImageProcessingService],
+})
+export class ImageProcessingModule {}
